+++ src/main/scala/mesosphere/marathon/core/appinfo/AppInfo.scala
@@ -14,7 +14,8 @@ case class AppInfo(
   maybeTasks: Option[Seq[EnrichedTask]] = None,
   maybeCounts: Option[TaskCounts] = None,
   maybeDeployments: Option[Seq[Identifiable]] = None,
-  maybeLastTaskFailure: Option[TaskFailure] = None)
+  maybeLastTaskFailure: Option[TaskFailure] = None,
+  maybeTaskStats: Option[TaskStatsByVersion] = None)
 
 object AppInfo {
   sealed trait Embed
@@ -23,5 +24,6 @@ object AppInfo {
     case object Deployments extends Embed
     case object Counts extends Embed
     case object LastTaskFailure extends Embed
+    case object TaskStats extends Embed
   }
 }